fix: integrate calendar to the new structure
